requirements.txt
scikit-learn==0.21.3
pandas==0.25.2
nltk==3.4.5
torch==1.5.0  --find-links https://download.pytorch.org/whl/torch_stable.html
cudatoolkit==10.1
transformers==2.8.0
tqdm==4.41.1
torchtext==0.5.0
BeautifulSoup4==4.8.0
langdetect==1.0.8
pyenchant==3.1.1
